backend/src/pyproject.toml
[tool.poetry]
name = "kwai"
version = "0.1.0"
description = "KWAI Sport Club Management - API"
authors = ["Franky Braem <franky.braem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
mysql-connector-python = "^8.0.31"
sql-smith = "^1.0.1"
pyisemail = "^2.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
loguru = "^0.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pylint = "^2.15.9"
black = {extras = ["d"], version = "^22.12.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
